Update scripts to allow quiet configuration for scripts

+++ .scripts/_env.bash
@@ -183,6 +183,7 @@ _do_scripts_env=(
 	TMPDIR="${_TMPDIR}"
 )
 
+_do_scripts_do_quiet="${pallur_do_scripts_do_quiet:-true}"
 _do_scripts_env_quiet="${pallur_do_scripts_env_quiet:-true}"
 
 while read _do_scripts_env_var ; do
@@ -213,15 +214,19 @@ fi
 function _do_exec () {
 	test "${#}" -ge 1
 	echo "[ii] executing \`${@:1}\`..." >&2
-	_do_exec_log="$( basename -- "${_workbench}" )--$( tr -d '\n' <<<"${*}" | md5sum -t | tr -d ' \n-' )--$( date '+%Y-%m-%d-%H-%M-%S-%N' ).log"
-	if test -e "${_temporary}" ; then
-		_do_exec_log="${_temporary}/${_do_exec_log}"
-	elif test -e "${_TMPDIR}" ; then
-		_do_exec_log="${_TMPDIR}/${_do_exec_log}"
-	elif test -e /tmp ; then
-		_do_exec_log="/tmp/${_do_exec_log}"
+	if test "${_do_scripts_do_quiet}" == true ; then
+		_do_exec_log="$( basename -- "${_workbench}" )--$( tr -d '\n' <<<"${*}" | md5sum -t | tr -d ' \n-' )--$( date '+%Y-%m-%d-%H-%M-%S-%N' ).log"
+		if test -e "${_temporary}" ; then
+			_do_exec_log="${_temporary}/${_do_exec_log}"
+		elif test -e "${_TMPDIR}" ; then
+			_do_exec_log="${_TMPDIR}/${_do_exec_log}"
+		elif test -e /tmp ; then
+			_do_exec_log="/tmp/${_do_exec_log}"
+		else
+			false
+		fi
 	else
-		false
+		_do_exec_log=/dev/stderr
 	fi
 	_outcome=0
 	setarch i686 --32bit --3gb -- \
@@ -229,12 +234,16 @@ function _do_exec () {
 	env -i "${_do_scripts_env[@]}" "${@}" </dev/null >"${_do_exec_log}" 2>&1 \
 	|| _outcome="${?}"
 	if test "${_outcome}" -ne 0 ; then
-		echo "[ee] failed with ${_outcome}; log available at \`${_do_exec_log}\`" >&2
-		tail -n 20 -- "${_do_exec_log}" | sed -u -r -e 's!^.*$![  ] &!g' >&2
-		echo "[--]" >&2
+		if test "${_do_scripts_do_quiet}" == true ; then
+			echo "[ee] failed with ${_outcome}; log available at \`${_do_exec_log}\`" >&2
+			tail -n 20 -- "${_do_exec_log}" | sed -u -r -e 's!^.*$![  ] &!g' >&2
+			echo "[--]" >&2
+		fi
 		return "${_outcome}"
 	else
-		rm -- "${_do_exec_log}"
+		if test "${_do_scripts_do_quiet}" == true ; then
+			rm -- "${_do_exec_log}"
+		fi
 		return 0
 	fi
 }
@@ -242,15 +251,19 @@ function _do_exec () {
 function _do_bash () {
 	test "${#}" -ge 1
 	echo "[ii] executing \`${@:1}\`..." >&2
-	_do_exec_log="$( basename -- "${_workbench}" )--$( tr -d '\n' <<<"${*}" | md5sum -t | tr -d ' \n-' )--$( date '+%Y-%m-%d-%H-%M-%S-%N' ).log"
-	if test -e "${_temporary}" ; then
-		_do_exec_log="${_temporary}/${_do_exec_log}"
-	elif test -e "${_TMPDIR}" ; then
-		_do_exec_log="${_TMPDIR}/${_do_exec_log}"
-	elif test -e /tmp ; then
-		_do_exec_log="/tmp/${_do_exec_log}"
+	if test "${_do_scripts_do_quiet}" == true ; then
+		_do_exec_log="$( basename -- "${_workbench}" )--$( tr -d '\n' <<<"${*}" | md5sum -t | tr -d ' \n-' )--$( date '+%Y-%m-%d-%H-%M-%S-%N' ).log"
+		if test -e "${_temporary}" ; then
+			_do_exec_log="${_temporary}/${_do_exec_log}"
+		elif test -e "${_TMPDIR}" ; then
+			_do_exec_log="${_TMPDIR}/${_do_exec_log}"
+		elif test -e /tmp ; then
+			_do_exec_log="/tmp/${_do_exec_log}"
+		else
+			false
+		fi
 	else
-		false
+		_do_exec_log=/dev/stderr
 	fi
 	_outcome=0
 	setarch i686 --32bit --3gb \
@@ -258,12 +271,16 @@ function _do_bash () {
 	env -i "${_do_scripts_env[@]}" BASH_ENV="${_scripts}/_env.bash" bash -- "${@}" </dev/null >"${_do_exec_log}" 2>&1 \
 	|| _outcome="${?}"
 	if test "${_outcome}" -ne 0 ; then
-		echo "[ee] failed with ${_outcome}; log available at \`${_do_exec_log}\`" >&2
-		tail -n 20 -- "${_do_exec_log}" | sed -u -r -e 's!^.*$![  ] &!g' >&2
-		echo "[--]" >&2
+		if test "${_do_scripts_do_quiet}" == true ; then
+			echo "[ee] failed with ${_outcome}; log available at \`${_do_exec_log}\`" >&2
+			tail -n 20 -- "${_do_exec_log}" | sed -u -r -e 's!^.*$![  ] &!g' >&2
+			echo "[--]" >&2
+		fi
 		return "${_outcome}"
 	else
-		rm -- "${_do_exec_log}"
+		if test "${_do_scripts_do_quiet}" == true ; then
+			rm -- "${_do_exec_log}"
+		fi
 		return 0
 	fi
 }
